refactor(db-products.js): simplify database queries by removing try-catch blocks and handling errors at a higher level refactor(validations/products.js): move validation logic to a dedicated validation module for better organization chore(products.js): remove unused user authorization function to streamline code
